Added ThemeWrapper component
