+++ components/AbbrTag/index.js
@@ -1,0 +1,5 @@
+const AbbrTag = props => (
+  <abbr title={props.mark.title}>{props.children}</abbr>
+)
+
+export default AbbrTag
