+++ package.json
@@ -16,6 +16,7 @@
     "es6-promise": "^4.2.8",
     "exenv": "^1.2.2",
     "isomorphic-fetch": "^2.2.1",
+    "lodash.merge": "^4.6.2",
     "next": "^9.2",
     "next-cookies": "^2.0.1",
     "node-sass": "^4.12.0",
