+++ themes/victorian/serializers.js
@@ -1,5 +1,11 @@
+const TestTag = props => (
+  <strong style={{'color': 'red'}}>{props.children}</strong>
+)
+
 const serializers = {
-	'hello': 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec porta sed ante accumsan gravida. Proin sollicitudin tincidunt nulla et ultrices. Cras tristique lacus id luctus dapibus. Vestibulum in leo efficitur, finibus magna ut, elementum nibh. Duis nibh quam, rhoncus nec risus et, commodo vehicula urna. Proin finibus bibendum mi, ut facilisis diam scelerisque ac. Duis eu auctor dui. Nam interdum condimentum mauris, id tincidunt nisl mattis non. Donec urna augue, imperdiet eu feugiat a, commodo ut leo. Vestibulum venenatis ligula a congue maximus. Cras vel pharetra dui. Sed bibendum ornare blandit. Donec tincidunt ligula ut lorem fringilla, vel sollicitudin ligula faucibus. Quisque rhoncus vestibulum tellus eget placerat. Donec fermentum, dolor et dictum faucibus, sapien tellus faucibus sem, sed fermentum dui nunc non felis. Etiam dapibus aliquam lacus ac lobortis. Nunc ornare odio elit, in maximus massa accumsan quis. Aliquam posuere tempus lobortis. Vestibulum laoreet tempor velit, a placerat est tincidunt sed. In at erat tincidunt, euismod orci vel, fermentum ipsum. Morbi est nibh, molestie sit amet pharetra non, bibendum sed arcu. Suspendisse non mauris tempor, lacinia turpis non, interdum est. Ut nisl nulla, porttitor sit amet rhoncus fermentum, blandit a ipsum. Morbi id venenatis nulla. Sed pretium vitae ante vitae fermentum. Duis accumsan leo ac est scelerisque luctus. Phasellus urna sapien, vestibulum nec mattis eu, eleifend ut arcu. Suspendisse venenatis congue dui, eu pulvinar neque euismod tincidunt. Quisque luctus lacinia neque, non cursus magna egestas vitae. Mauris volutpat, lorem nec rutrum mattis, metus enim facilisis sem, id blandit felis massa ut nisi. Nunc sed pretium augue. Vivamus ac metus ut turpis lobortis interdum in ac ligula. Nulla facilisi. Donec volutpat tempus augue non dapibus. Donec quis lobortis lectus. Donec vitae elit eu augue fermentum scelerisque. Maecenas ullamcorper arcu a arcu ornare, non sodales diam semper. In ut ante vel mi euismod congue. Curabitur dignissim mollis tellus at dapibus. Curabitur non urna nec mauris faucibus consequat. Etiam magna risus, feugiat quis imperdiet dignissim, mollis ac urna. Aenean sodales at turpis eget accumsan. Maecenas accumsan dui pharetra vestibulum dapibus. Suspendisse faucibus porta consectetur. Sed vulputate maximus iaculis. Fusce eleifend suscipit imperdiet. Phasellus mattis mattis est, suscipit sagittis massa mollis quis. Praesent vel molestie felis. Suspendisse feugiat pretium mauris, nec varius velit viverra ac. Sed id imperdiet metus. Sed pellentesque ex erat, at posuere mauris malesuada sit amet. Praesent vulputate nulla lectus, sit amet pulvinar mauris molestie porta. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Aliquam id ex pulvinar, semper nunc vel, tempor lectus. Nullam vehicula nisi eu odio porta molestie eu eget sem. Donec dictum ipsum a massa iaculis, id commodo tortor venenatis. Phasellus imperdiet pellentesque magna, in molestie justo. Maecenas sed lorem lacus. Ut porttitor, ipsum vel rhoncus tristique, lectus nisi ornare sem, vitae suscipit justo ipsum vitae velit. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Sed ac felis consectetur, posuere libero vel, consequat justo. Vivamus tempor aliquam faucibus. Aenean a metus dictum, dignissim nulla quis, consequat risus. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Sed a lectus id sem tristique hendrerit eu sit amet massa. Etiam ultrices eros tortor, vitae laoreet nibh hendrerit vel. Aenean tincidunt diam sed sollicitudin feugiat. Nullam dapibus consequat velit, ut ullamcorper elit aliquam eu. Maecenas rutrum ligula sit amet enim venenatis sollicitudin. Pellentesque et neque elit. Quisque scelerisque turpis ac tortor tristique posuere. Phasellus ut ante vitae ligula tincidunt tempor sit amet quis metus. Mauris aliquam velit quam, eu suscipit urna ultrices sit amet. Fusce cursus dui neque, non suscipit ante cursus eu. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Cras molestie enim odio. Aliquam mauris diam, imperdiet quis risus sit amet, faucibus molestie lorem. In viverra mollis posuere. Pellentesque luctus auctor fringilla. Fusce vehicula rutrum risus, eu porttitor eros aliquam placerat. Praesent eget pretium nunc, a hendrerit metus. Donec dignissim ullamcorper orci pretium molestie. Vestibulum feugiat felis consectetur vehicula finibus. Cras magna nisl, pulvinar nec libero tristique, commodo lacinia libero. Etiam viverra erat id est venenatis imperdiet. Vestibulum vitae dui tempus arcu pulvinar maximus in sit amet urna. Ut ultrices, erat vitae blandit maximus, lorem magna vehicula ante, eget consectetur odio nunc non ligula. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Integer bibendum, magna id euismod imperdiet, velit neque lacinia lorem sed. '
+	marks: {
+    abbrTag: TestTag
+  }
 }
 
 export default serializers
